requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
sentence-transformers
faiss-cpu
langchain-community
pdfplumber
langchain